Fix memory issues with Db
There are currently two memory issue with Db. The first comes from
hanle.syncdbs(), allowing you to get multiple Dbs that point to the same
backing Db.

With this, you could view the servers on one Db, and call .set_servers()
on the other. This would invalidate the reference used to view the
servers originally.

The second issue is that updating a db invalidates any pkgcache pointer
out there.

The fix here is to simply have Db mutably borrow the handle. So now the
type DbMut exists and derefs to Db. As well at mut versions of many
existing Db functions. Do note that a MutDb is not needed to set usage,
as that does not hold a reference to any parts of the DB.

Although this fixes the problem, this makes the API quite annoying to
use. Say you want to create a Db and add some servers. You can do that,
but as long as the DbMut exists, your handle will be mutably
borrowed. You could drop the DbMut, then get a regular Db via
handle.syncdbs(), but that's annoying.

Finding a nicer API for this is TODO.