Ensure loaded packages are only freed when they should
Most of the time, a package comes from the database and will be freed
when the database is unregistered.

However packages loaded from a file need to be freed manually by the
user. So our rust code needs to make sure the loaded package can not go
out of scope before it is done being use.

There is one exception to this, alpm_pkg_add. When a package is added to
the transaction. The ownership is effectivley moved into alpm. And alpm
will then free that package when it's done with it.

Do handle this, three different types of packages are needed.

LoadedPackage: a package loaded from a file
Package: a package loaded from the database
Pkg: a deref target for both above packages

For the special alpm_pkg_add function, packages will not run their
destructor when moved into the function, unless the function errors.