Fix question callbacks taking reference to stack data
Perviously, when getting a pointer to a union field we did:

	let inner = &mut unsafe { (*self.inner).replace }

Due to the &mut being outside the unsafe block this actually takes
a reference to a temporary variables instead of the union filed.

It effectively looks like:

	let tmp = unsafe { (*self.inner).replace };
	let inner = &mut tmp //reference to stack data

The fix is to just move the &mut into the safe block, making rust not
create a temporary variable and actually reference the union.