Store Db as NonNull
This doesn't really give us any benifit now but the goal it to
eventually remove the alpm reference from the struct and use the db for
the nullptr optimization.

Though looking at the asm this does knock off one instruction where the
program explicitly sets the Option<> tag to 0 as it can now just use the
pointer as the tag.