Encapsulate return type with `Option` enum instead of panicking
It's possible for libalpm to return valid errors without a pointer.

Therefore change the return type for the `data()` function in the
`PrepareError` and `CommitError` structs to be encapsulated within an
`Option` enum in order to prevent a fatal panic, instead returning `None`
when an error is unhandled and thus contains no data.